requirements.txt
Flask>=3.0.0
numpy<2.0.0
tensorflow>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
Pillow>=10.0.0
